Carbon Processor. Add KMP support
