+++ composites-carbon/processor/src/commonMain/kotlin/ru/ldralighieri/composites/carbon/processor/model/CarbonRouteData.kt
