+++ composites-carbon/processor/src/commonMain/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
