--- composites-carbon/processor/src/main/kotlin/ru/ldralighieri/composites/carbon/processor/model/MemberNames.kt
@@ -1,6 +1,0 @@
-package ru.ldralighieri.composites.carbon.processor.model
-
-import com.squareup.kotlinpoet.MemberName
-
-internal val navArgumentMemberName = MemberName("androidx.navigation", "navArgument")
-internal val navDeepLinkMemberName = MemberName("androidx.navigation", "navDeepLink")
